Add call filtering in event callback to avoid memory bloat.

+++ ext/fiber/profiler/time.c
@@ -14,7 +14,7 @@ void Fiber_Profiler_Time_elapsed(const struct timespec* start, const struct time
 	}
 }
 
-float Fiber_Profiler_Time_duration(const struct timespec *duration)
+double Fiber_Profiler_Time_duration(const struct timespec *duration)
 {
 	return duration->tv_sec + duration->tv_nsec / 1000000000.0;
 }
@@ -23,11 +23,11 @@ void Fiber_Profiler_Time_current(struct timespec *time) {
 	clock_gettime(CLOCK_MONOTONIC, time);
 }
 
-float Fiber_Profiler_Time_proportion(const struct timespec *duration, const struct timespec *total_duration) {
+double Fiber_Profiler_Time_proportion(const struct timespec *duration, const struct timespec *total_duration) {
 	return Fiber_Profiler_Time_duration(duration) / Fiber_Profiler_Time_duration(total_duration);
 }
 
-float Fiber_Profiler_Time_delta(const struct timespec *start, const struct timespec *stop) {
+double Fiber_Profiler_Time_delta(const struct timespec *start, const struct timespec *stop) {
 	struct timespec duration;
 	Fiber_Profiler_Time_elapsed(start, stop, &duration);
