+++ ext/fiber/profiler/time.h
@@ -7,11 +7,11 @@
 #include <time.h>
 
 void Fiber_Profiler_Time_elapsed(const struct timespec* start, const struct timespec* stop, struct timespec *duration);
-float Fiber_Profiler_Time_duration(const struct timespec *duration);
+double Fiber_Profiler_Time_duration(const struct timespec *duration);
 void Fiber_Profiler_Time_current(struct timespec *time);
 
-float Fiber_Profiler_Time_delta(const struct timespec *start, const struct timespec *stop);
-float Fiber_Profiler_Time_proportion(const struct timespec *duration, const struct timespec *total_duration);
+double Fiber_Profiler_Time_delta(const struct timespec *start, const struct timespec *stop);
+double Fiber_Profiler_Time_proportion(const struct timespec *duration, const struct timespec *total_duration);
 
 #define Fiber_Profiler_TIME_PRINTF_TIMESPEC "%.3g"
 #define Fiber_Profiler_TIME_PRINTF_TIMESPEC_ARGUMENTS(ts) ((double)(ts).tv_sec + (ts).tv_nsec / 1e9)
